+++ gamelibrary.h
@@ -20,11 +20,12 @@ public:
     void init(const QString& directory);
     void setupLibrary(bool force = false);
     void setupLibrary(QString directory, bool force);
-    void process(QByteArray qbyteArray);
+    void setupDatabase(QByteArray qbyteArray);
     bool load(QString filepath);
     bool save(QString filepath);
     void dump();
 
+    QString jsonFile;
     QString baseDirectory;
     QMap<QString, LibraryEntry*> library;
     QMap<QString, TitleInfo*> database;